fix density of neutral helium lookup table test
